Add venv support to Python template
Closes #24


Former-commit-id: c5d0172441df8d20ec96ebbae2b8d259bfae1ba5

+++ 2023/day-25/build-ci.sh
@@ -1,4 +1,7 @@
 #!/bin/sh
-pypy3 -m venv venv
-. venv/bin/activate
-pypy3 -m pip install -r requirements.txt
+if [ -f requirements.txt ]; then
+    pypy3 -m venv venv
+    # shellcheck disable=SC1091
+    . venv/bin/activate
+    pypy3 -m pip install -r requirements.txt
+fi
